Add support for use with PHPUnit Phar
When PHPUnit is packaged as a Phar, certain classes are prefixed to prevent conflicts with project dependencies which may use the same files, but in a different version.

For this package, only one referenced class is impacted by this.

The small fix now applied allows for using this package both when running PHPUnit installed via Composer, as well as running Composer installed as a Phar.