README: clarify code examples
To use the functionality from the library, users can choose one of two methods to use it.

The existing code example in the readme was confusing as it used both methods in the same example, while at any time, only one method should be used.

This splits out the example into the two distinct usages.

Includes:
* Update namespace in code samples to make it clear that the user namespace is expected.
* Update test class name in code samples to make it clear this is an example.
* Remove `strict_types` declaration.
    While this may just be my opinion, `strict_types` should not be used in test files as it can block the testing of unhappy paths.