GH Actions: adjust the test workflow
... to:
* Run the availability tests against all PHP versions on which the package can be installed with `prefer-stable`.
* Run the availability tests as well as the _real_ tests against `lowest` and `stable` for PHPUnit 9 on PHP 7.3 - 8.0.

The test selection is done automatically via the `phpunit.xml.dist` file as adjusted in the previous commit.

As for the running of the _real_ tests against the `lowest` supported PHPUnit 9 version, this requires some tweaking to the workflow script.
As PHPUnit didn't always have an explicit requirement for PHP, running with `prefer-lowest` on PHP 7.3 - 8.0 with the current PHPUnit version constraints would install PHPUnit 4.8.
So, for the `prefer-lowest` runs, we need to make sure that `lowest` is correctly interpreted as "lowest" within the PHPUnit 9.x range.

Also:
* As the coding standards package has a minimum PHP requirement of PHP 7.2, it would block the `composer install` for lower PHP versions to run the tests, so removing the package for this workflow.