Allow the polyfill to kick in with PHPUnit 8.x
PHPUnit 8.x deprecated the `assertArraySubset()` functionality, PHPUnit 9.x removed it.

Up to now, this package would only "kick in" in combination with PHPUnit 9.x to polyfill the functionality.

The combined changes in this commit will allow for this package to also "kick in" in combination with PHPUnit 8.x, preventing the PHPUnit native deprecation for using `assertArraySubset()` from being thrown.

This commit also includes CI changes to safeguard the correct functioning of the code when used with PHPUnit 8.x.

Notes:
* The main change is in the autoloader, most other files only contain changed annotations.
* The only functional change needed was a work-around for the PHPUnit `InvalidArgumentException`, which was only introduced in PHPUnit 8.4.0.
* This commit does change the minimum PHP version code within the actual polyfill has to be compatible with.
    Previously, the code has to be compatible with PHP 7.3 and higher (minimum supported PHP version for PHPUnit 9.x).
    Now, the code has to be compatible with PHP **7.2** and higher (minimum supported PHP version for PHPUnit 8.x).

Fixes 50